Fixes #31. DRY up scheduled call create and cancel. rake tasks for production.
