--- app/controllers/occurences_controller.rb
@@ -1,5 +1,0 @@
-class OccurencesController < ApplicationController
-  def create
-    Occurence.make_summit_request
-  end
-end
