+++ app/views/admin/schedules/_new.html.erb
@@ -76,5 +76,5 @@
 
    <% end %>
   <button type="button" data-dismiss="modal" id="add-new-schedule-close" class="btn btn-default md-close">Cancel</button>
-  <%= f.submit "Submit", 'data-dismiss': 'modal', class: 'btn btn-primary md-close'  %>
+  <%= f.submit "Submit", 'data-dismiss': 'modal', class: 'btn btn-primary md-close', id: 'new-schedule-submit'  %>
 <% end %>
