+++ app/views/admin/schedules/index.html.erb
@@ -11,7 +11,7 @@
           <div class="widget widget-fullwidth widget-small">
             <div class="widget-head">
               <div class="tools">
-                <a data-modal="add_new_schedule" class="md-trigger">
+                <a data-modal="add_new_schedule" class="md-trigger" id="add-schedule-button">
                   <span class="icon s7-add-user"></span>
                 </a>
                 </div>
