+++ lib/tasks/occurences.rake
@@ -4,10 +4,9 @@ namespace :occurences do
     occurences_without_scheduled_calls = Occurence.includes(:scheduled_calls).where( :scheduled_calls => { :occurence_id => nil } )
 
     occurences_without_scheduled_calls.each do |occurence|
-      occurence.make_summit_request
+      occurence.create_scheduled_call
     end
   end
-
 
   desc "Creates next 3 occurences of schedule"
   task next_3_occurences: :environment do
@@ -22,8 +21,9 @@ namespace :occurences do
     create_next_3_occurences(schedules_needing_occurences)
   end
 
+
   desc "Deletes past occurences"
-  task delete_past_occurences: :environment do
+  task delete_past_occurences_and_scheduled_calls: :environment do
     now = DateTime.now
     occurences_in_the_past = Occurence.where("time < ?", now)
     occurences_in_the_past.each do |occ|
@@ -33,24 +33,9 @@ namespace :occurences do
   end
 
 
-  desc "Deletes cancelled scheduled_calls"
-  task delete_cancelled_scheduled_calls: :environment do
-    ScheduledCall.where(cancelled: true).delete_all
-  end
-
-  desc "Deletes orphaned occurences and scheduled calls"
-  task delete_orphaned_occurences_and_schedules: :environment do
-    include SummitRequests
+  desc "Deletes orphaned occurences"
+  task delete_orphaned_occurences: :environment do
     orphaned_occurences = Occurence.where(["schedule_id NOT IN (?)", Schedule.select("id")]).delete_all
-    orphaned_scheduled_calls = ScheduledCall.where(["schedule_id NOT IN (?)", Schedule.select("id")]).where(cancelled: true).delete_all
-
-    uncancelled_calls = ScheduledCall.where(["schedule_id NOT IN (?)", Schedule.select("id")]).where(cancelled: false)
-
-    uncancelled_calls.each do |call|
-      SummitRequests.cancel_scheduled_call(call)
-      call.cancelled = true
-      call.save!
-    end
   end
 end
 
@@ -69,14 +54,15 @@ def create_next_3_occurences(schedules)
     end
 
     occurence_array = Montrose.every(:week).on(frequency.repeat_days).at(frequency.format_time).starts(day_after_last_occurence).take(3)
-    a = []
     occurence_array.each do |occurence|
-      a << Occurence.new(time: occurence.utc, schedule: schedule)
-    end
-
-    if a.each(&:save!)
-      schedule.last_occurence_date = get_last_occurence_date(a)
-      schedule.save!
+      occ = Occurence.new(time: occurence.utc, schedule: schedule)
+      if occ.save!
+        scheduled_call = occ.create_scheduled_call
+        schedule.last_occurence_date = get_last_occurence_date(schedule.occurences)
+        schedule.save!
+      else
+        Rails.logger.error = "Error creating occurence and scheduled_call for schedule with id " + schedule.id
+      end
     end
   end
 end
@@ -86,3 +72,6 @@ def get_last_occurence_date(occurences)
   last_occurence = occurences.sort_by(&:time).last
   last_occurence.time.to_date
 end
+
+
+
